requirements.txt
Pillow==5.3.0
pygame==1.9.4
mss==3.3.2